Fix unlimited FPS (real)

+++ src/core/Constants.java
@@ -50,6 +50,6 @@ public final class Constants {
 
 
     public static final float PLAYER_MAX_SPEED = 5;
-    public static final float SCALING_FACTOR() { return Main.config.UNLIMITED_FPS ? 60f / Main.getRealFPS() : 60f / Main.config.FRAMES_PER_SECOND; }
+    public static final float SCALING_FACTOR() { return Main.config.UNLIMITED_FPS ? 60f / (float) Main.getRealFPS() : 60f / Main.config.FRAMES_PER_SECOND; }
 
 }
